Started an initial solution which only partially works and added a correction forum post to alert the admin team to fix the initial Java code
I have started a solution which works for some of the cases, but not all of them. I believe the issue is that when the code tries to determine which Pawn can legally be promoted, it only determines this by checking if there are any enemy pieces in front and blocking it. However, the code does not account for pins. For example, if a Pawn is blocking its King from being captured by a Bishop, it should not legally be allowed to move, as it would be putting its King into check. However, the code does not account for this. I shall fix this in the next commit.

Also, while testing, I found that the initial Java code Hackerrank provides causes a Runtime error (Specifically a  java.util.NoSuchElementException Runtime error). I posted on the forums and used the "Suggest edit" form to bring this to the attention of the admin team and anyone attempting this using Java (Details in CorrectionPost.md)